pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "cullenetal"
version = "0.1.0"
description = "Reproduction of 'Grokking as a Phase Transition between Competing Basins' (Cullen et al., 2026)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "scipy>=1.10",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "devinterp>=0.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
